requirements.txt
wtforms==3.0.1
sqlalchemy==1.4.31
pillow==9.1.0
flask==2.0.2
flask-wtf==1.0.0
flask-login==0.5.0
gunicorn==20.1.0
Werkzeug==2.0.2
Jinja2==3.0.3